+++ spec/fixtures/cookbooks/oneview_test_api300_synergy/recipes/fabric_set_reserved_vlan_range.rb
@@ -1,4 +1,8 @@
-# (c) Copyright 2016-2017 Hewlett Packard Enterprise Development LP
+#
+# Cookbook Name:: oneview_test_api300_synergy
+# Recipe:: ethernet_network_create
+#
+# (c) Copyright 2016 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -8,19 +12,13 @@
 # under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
+#
 
-OneviewCookbook::ResourceBaseProperties.load(self)
-
-default_action :add
-
-action :add do
-  OneviewCookbook::Helper.do_resource_action(self, :SANManager, :add_or_edit)
-end
-
-action :add_if_missing do
-  OneviewCookbook::Helper.do_resource_action(self, :SANManager, :add_if_missing)
-end
-
-action :remove do
-  OneviewCookbook::Helper.do_resource_action(self, :SANManager, :remove)
+oneview_fabric 'Fabric1' do
+  client node['oneview_test']['client']
+  reserved_vlan_range(
+    'start' => 1200,
+    'length' => 120
+  )
+  action :set_reserved_vlan_range
 end