+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   - fcoe_network support for API300
   - network_set support for API300
   - datacenter support for API300
+  - fabric new resource support for API300 Synergy
+  - san_manager support for API300
   - rack support for API300
 
 ### 1.1.0